+++ _pkg/tracker/ua/ua_blacklist.go
@@ -584,6 +584,7 @@ var UserAgentBlacklist = []string{
 	"inreachapplication",
 	"insider-gaming",
 	"insomnia",
+	"inspection",
 	"installatron",
 	"instapaper",
 	"integration",
@@ -1030,6 +1031,7 @@ var UserAgentBlacklist = []string{
 	"safbar",
 	"safeassign",
 	"sandbox",
+	"sapphire",
 	"saws",
 	"scalingo",
 	"scamadviserexternalhit",
@@ -1393,5 +1395,6 @@ var UserAgentRegexBlacklist = []*regexp.Regexp{
 	regexp.MustCompile("assert\\(base64_decode\\("),
 	regexp.MustCompile("case *when.*then.*else.*end"),
 	regexp.MustCompile("expr [0-9]* - [0-9]*"),
+	regexp.MustCompile("firefox/(\\d+\\.\\d+\\.\\d+)"),
 	regexp.MustCompile("iif *(.*,.*,.*)"),
 }